First probably broken attempt to refactor QbWriter into some smaller individual files.
I have no idea if the tests work at all.
