+++ src/csvcubed/writers/qbwriter.py
@@ -7,46 +7,28 @@ Output writer for CSV-qb
 import itertools
 import json
 import logging
-import re
-from dataclasses import field
 from pathlib import Path
-from typing import Tuple, Dict, Any, List, Iterable, Set
+from typing import Tuple, Dict, Any, Iterable
 
 import pandas as pd
 
-from csvcubedmodels.rdf.dependency import RdfGraphDependency
-from csvcubedmodels import rdf
-from csvcubedmodels.rdf import skos, rdfs
-from csvcubedmodels.rdf.resource import (
-    ExistingResourceWithLiteral,
-    Resource,
-    ExistingResource,
-    maybe_existing_resource,
-)
-
-from csvcubed.models.rdf import prov
 from csvcubed.models.cube import *
 from csvcubed.utils.uri import (
-    get_last_uri_part,
     csvw_column_name_safe,
-    get_data_type_uri_from_str,
-    uri_safe,
 )
 from csvcubed.utils.csvw import get_dependent_local_files
 from csvcubed.utils.qb.cube import (
     get_columns_of_dsd_type,
     QbColumnarDsdType,
 )
-from csvcubed.utils.dict import rdf_resource_to_json_ld
 from csvcubed.utils.qb.standardise import (
     convert_data_values_to_uri_safe_values,
     ensure_int_columns_are_ints,
 )
 from csvcubed.utils.file import copy_files_to_directory_with_structure
+from .dsdhelper import QbDsdHelper
 from .skoscodelistwriter import SkosCodeListWriter
-from .urihelpers.skoscodelist import SkosCodeListNewUriHelper
-from .urihelpers.qbcube import QbCubeNewUriHelper
-from .urihelpers.skoscodelistconstants import SCHEMA_URI_IDENTIFIER
+from .urihelpers.cubenew import QbUriHelper
 from .writerbase import WriterBase
 from ..models.cube import (
     QbAttribute,
@@ -56,12 +38,7 @@ from ..models.cube import (
     QbMultiUnits,
     QbAttributeLiteral,
 )
-from csvcubed.utils.version import get_csvcubed_version_uri
-from csvcubed.models.rdf.qbdatasetincatalog import QbDataSetInCatalog
 from ..models.cube.qb.components.codelist import NewQbCodeListInCsvW
-from ..models.rdf.newattributevalueresource import NewAttributeValueResource
-from ..models.rdf.newunitresource import NewUnitResource
-from ..models.cube.qb.components.arbitraryrdf import RdfSerialisationHint
 from ..utils.qb.validation.observations import get_observation_status_columns
 
 
@@ -75,34 +52,12 @@ class QbWriter(WriterBase):
     cube: QbCube
     csv_file_name: str = field(init=False)
     raise_missing_uri_safe_value_exceptions: bool = field(default=True, repr=False)
-    _new_uri_helper: QbCubeNewUriHelper = field(init=False)
+    _uris: QbUriHelper = field(init=False)
+    _dsd: QbDsdHelper = field(init=False)
 
     @property
     def csv_metadata_file_name(self) -> str:
         return f"{self.csv_file_name}-metadata.json"
-
-    @property
-    def is_cube_in_pivoted_shape(self) -> bool:
-        obs_val_columns = get_columns_of_dsd_type(self.cube, QbObservationValue)
-
-        all_pivoted = True
-        all_standard_shape = True
-        for obs_val_col in obs_val_columns:
-            all_pivoted = (
-                all_pivoted
-                and obs_val_col.structural_definition.is_pivoted_shape_observation
-            )
-            all_standard_shape = (
-                all_standard_shape
-                and not obs_val_col.structural_definition.is_pivoted_shape_observation
-            )
-
-        if all_pivoted:
-            return True
-        elif all_standard_shape:
-            return False
-        else:
-            raise TypeError("The cube cannot be in both standard and pivoted shape")
 
     def __post_init__(self):
         self.csv_file_name = f"{self.cube.metadata.uri_safe_identifier}.csv"
@@ -111,7 +66,8 @@ class QbWriter(WriterBase):
             QbWriter.__name__,
             self.csv_file_name,
         )
-        self._new_uri_helper = QbCubeNewUriHelper(self.cube)
+        self._uris = QbUriHelper(self.cube)
+        self._dsd = QbDsdHelper(self.cube, self._uris)
 
     def write(self, output_folder: Path):
         # Map all labels to their corresponding URI-safe-values, where possible.
@@ -143,7 +99,7 @@ class QbWriter(WriterBase):
                     "columns": self._generate_csvw_columns_for_cube(),
                     "foreignKeys": self._generate_foreign_keys_for_cube(),
                     "primaryKey": self._get_primary_key_columns(),
-                    "aboutUrl": self._get_about_url(),
+                    "aboutUrl": self._uris.get_about_url(),
                 },
             }
         ]
@@ -154,9 +110,9 @@ class QbWriter(WriterBase):
 
         csvw_metadata = {
             "@context": "http://www.w3.org/ns/csvw",
-            "@id": self._new_uri_helper.get_dataset_uri(),
+            "@id": self._uris.get_dataset_uri(),
             "tables": tables,
-            "rdfs:seeAlso": self._get_additional_rdf_metadata(),
+            "rdfs:seeAlso": self._dsd.generate_data_structure_definitions(),
         }
 
         metadata_json_output_path = output_folder / self.csv_metadata_file_name
@@ -168,23 +124,6 @@ class QbWriter(WriterBase):
             csv_output_file_path = output_folder / self.csv_file_name
             _logger.debug("Writing CSV to %s", csv_output_file_path)
             self.cube.data.to_csv(csv_output_file_path, index=False)
-
-    def _get_additional_rdf_metadata(self) -> List[dict]:
-        """
-        :return: the additional RDF metadata to be serialised in the CSV-W.
-        """
-        see_also = rdf_resource_to_json_ld(self._generate_qb_dataset_dsd_definitions())
-
-        for dependencies in self._get_rdf_file_dependencies():
-            see_also += rdf_resource_to_json_ld(dependencies)
-
-        for attribute_value in self._get_new_attribute_value_resources():
-            see_also += rdf_resource_to_json_ld(attribute_value)
-
-        for unit in self._get_new_unit_resources():
-            see_also += rdf_resource_to_json_ld(unit)
-
-        return see_also
 
     def _output_new_code_list_csvws(self, output_folder: Path) -> None:
         for column in get_columns_of_dsd_type(self.cube, NewQbDimension):
@@ -346,7 +285,7 @@ class QbWriter(WriterBase):
                 "virtual": True,
                 "aboutUrl": observation_uri,
                 "propertyUrl": "qb:measureType",
-                "valueUrl": self._get_measure_uri(measure),
+                "valueUrl": self._uris.get_measure_uri(measure),
             }
         )
         # For each dimension in the cube, creates the `?obsUri ?dimUri ?valueUri` triple.
@@ -359,7 +298,7 @@ class QbWriter(WriterBase):
             (
                 property_url,
                 value_url,
-            ) = self._get_default_property_value_uris_for_dimension(dimension_col)
+            ) = self._uris.get_default_property_value_uris_for_dimension(dimension_col)
 
             if dimension_col.csv_column_uri_template is not None:
                 _logger.debug(
@@ -397,7 +336,7 @@ class QbWriter(WriterBase):
                 "virtual": True,
                 "aboutUrl": observation_uri,
                 "propertyUrl": "qb:dataSet",
-                "valueUrl": self._new_uri_helper.get_dataset_uri(),
+                "valueUrl": self._uris.get_dataset_uri(),
             }
         )
 
@@ -426,7 +365,7 @@ class QbWriter(WriterBase):
                 "name": "virt_slice_structure",
                 "virtual": True,
                 "propertyUrl": "qb:sliceStructure",
-                "valueUrl": self._new_uri_helper.get_slice_key_across_measures_uri(),
+                "valueUrl": self._uris.get_slice_key_across_measures_uri(),
             }
         )
 
@@ -475,7 +414,7 @@ class QbWriter(WriterBase):
                 "name": "virt_dataset",
                 "virtual": True,
                 "propertyUrl": "http://purl.org/linked-data/cube#dataSet",
-                "valueUrl": self._new_uri_helper.get_dataset_uri(),
+                "valueUrl": self._uris.get_dataset_uri(),
             },
         ]
         unit = obs_val.unit
@@ -486,504 +425,14 @@ class QbWriter(WriterBase):
                     "name": VIRT_UNIT_COLUMN_NAME,
                     "virtual": True,
                     "propertyUrl": "http://purl.org/linked-data/sdmx/2009/attribute#unitMeasure",
-                    "valueUrl": self._get_unit_uri(unit),
+                    "valueUrl": self._uris.get_unit_uri(unit),
                 }
             )
 
         return virtual_columns
 
-    def _get_qb_dataset_with_catalog_metadata(self) -> QbDataSetInCatalog:
-        qb_dataset_with_metadata = QbDataSetInCatalog(
-            self._new_uri_helper.get_dataset_uri()
-        )
-        self.cube.metadata.configure_dcat_dataset(qb_dataset_with_metadata)
-        return qb_dataset_with_metadata
-
-    def _generate_qb_dataset_dsd_definitions(self) -> QbDataSetInCatalog:
-        dataset = self._get_qb_dataset_with_catalog_metadata()
-
-        generation_activity = prov.Activity(self._new_uri_helper.get_activity_uri())
-        generation_activity.used = ExistingResource(get_csvcubed_version_uri())
-        dataset.was_generated_by = generation_activity
-
-        dataset.structure = rdf.qb.DataStructureDefinition(
-            self._new_uri_helper.get_structure_uri()
-        )
-        component_ordinal = 1
-        for column in self.cube.columns:
-            if isinstance(column, QbColumn):
-                component_specs_for_col = self._get_qb_component_specs_for_col(
-                    column.uri_safe_identifier, column.structural_definition
-                )
-                component_properties_for_col = [
-                    p for s in component_specs_for_col for p in s.componentProperties
-                ]
-                dataset.structure.componentProperties |= set(
-                    component_properties_for_col
-                )
-                for component in component_specs_for_col:
-                    component.order = component_ordinal
-                    component_ordinal += 1
-
-                dataset.structure.components |= set(component_specs_for_col)
-
-        if self.is_cube_in_pivoted_shape:
-            dataset.structure.sliceKey.add(self._get_cross_measures_slice_key())
-
-        return dataset
-
-    def _get_cross_measures_slice_key(self) -> rdf.qb.SliceKey:
-        # Setting up Slice Key for slices which range over measures.
-        _logger.debug("Setting the slice key across measures")
-
-        slice_key = rdf.qb.SliceKey(
-            self._new_uri_helper.get_slice_key_across_measures_uri()
-        )
-        _logger.debug("Slice key uri across measures is '%s'", slice_key.uri)
-
-        for dimension_column in get_columns_of_dsd_type(self.cube, QbDimension):
-            _logger.debug(
-                "Adding the component property for dimension column with title '%s' into the slice key",
-                dimension_column.csv_column_title,
-            )
-            dimension = dimension_column.structural_definition
-            dimension_uri: str
-            if isinstance(dimension, NewQbDimension):
-                _logger.debug("The dimension column is a new dimension")
-                dimension_uri = self._new_uri_helper.get_dimension_uri(
-                    dimension.uri_safe_identifier
-                )
-                _logger.debug(
-                    "The dimension column is a new dimension with uri '%s'",
-                    dimension_uri,
-                )
-            elif isinstance(dimension, ExistingQbDimension):
-                dimension_uri = dimension.dimension_uri
-                _logger.debug(
-                    "The dimension column is an existing dimension with uri '%s'",
-                    dimension_uri,
-                )
-            else:
-                raise TypeError(f"Unmatched QbDimension type {type(dimension)}")
-
-            slice_key.componentProperties.add(ExistingResource(dimension_uri))
-
-        _logger.debug(
-            "Added %d component properties to the slice key.",
-            len(slice_key.componentProperties),
-        )
-        return slice_key
-
-    def _get_qb_component_specs_for_col(
-        self, column_name_uri_safe: str, component: QbStructuralDefinition
-    ) -> Iterable[rdf.qb.ComponentSpecification]:
-        if isinstance(component, QbDimension):
-            return [
-                self._get_qb_dimension_specification(column_name_uri_safe, component)
-            ]
-        elif isinstance(component, QbAttribute):
-            return [
-                self._get_qb_attribute_specification(column_name_uri_safe, component)
-            ]
-        elif isinstance(component, QbMultiUnits):
-            return [self._get_qb_units_column_specification(column_name_uri_safe)]
-        elif isinstance(component, QbMultiMeasureDimension):
-            return self._get_qb_measure_dimension_specifications(component)
-        elif isinstance(component, QbObservationValue):
-            return self._get_qb_obs_val_specifications(component)
-        else:
-            raise TypeError(f"Unhandled component type {type(component)}")
-
-    def _get_obs_val_data_type(self) -> str:
-        observation_value_columns = get_columns_of_dsd_type(
-            self.cube, QbObservationValue
-        )
-        # Given the data shapes we accept as input, there should always be one (and only one) Observation Value
-        # column in a cube.
-        observation_value_column = observation_value_columns[0]
-        data_type = observation_value_column.structural_definition.data_type
-        return get_data_type_uri_from_str(data_type)
-
-    def _get_qb_units_column_specification(
-        self, column_name_uri_safe: str
-    ) -> rdf.qb.AttributeComponentSpecification:
-        component = rdf.qb.AttributeComponentSpecification(
-            self._new_uri_helper.get_component_uri(column_name_uri_safe)
-        )
-        component.componentRequired = True
-        component.attribute = ExistingResource(
-            "http://purl.org/linked-data/sdmx/2009/attribute#unitMeasure"
-        )
-        component.componentProperties.add(component.attribute)
-
-        _logger.debug(
-            "Generated units component %s.",
-            component.uri,
-        )
-
-        return component
-
-    def _get_qb_obs_val_specifications(
-        self, observation_value: QbObservationValue
-    ) -> List[rdf.qb.ComponentSpecification]:
-        specs: List[rdf.qb.ComponentSpecification] = [
-            # We always output the measure-dimension style of the QB spec.
-            # so each observation need to have a dimension specifying the measure type.
-            self._get_measure_type_dimension_component_spec()
-        ]
-
-        unit = observation_value.unit
-        if unit is not None:
-            specs.append(self._get_qb_units_column_specification("unit"))
-
-        if observation_value.is_pivoted_shape_observation:
-            assert observation_value.measure is not None
-            specs.append(
-                self._get_qb_measure_component_specification(observation_value.measure)
-            )
-
-        return specs
-
-    def _get_measure_type_dimension_component_spec(
-        self,
-    ) -> rdf.qb.DimensionComponentSpecification:
-        measure_dimension_spec = rdf.qb.DimensionComponentSpecification(
-            self._new_uri_helper.get_component_uri("measure-type")
-        )
-        measure_dimension_spec.dimension = ExistingResource(
-            "http://purl.org/linked-data/cube#measureType"
-        )
-        measure_dimension_spec.componentProperties.add(measure_dimension_spec.dimension)
-        return measure_dimension_spec
-
-    @staticmethod
-    def _get_unit_uri_safe_identifier(unit: QbUnit) -> str:
-        if isinstance(unit, ExistingQbUnit):
-            return get_last_uri_part(unit.unit_uri)
-        elif isinstance(unit, NewQbUnit):
-            return unit.uri_safe_identifier
-        else:
-            raise TypeError(f"Unhandled unit type {type(unit)}")
-
-    def _get_qb_measure_dimension_specifications(
-        self, measure_dimension: QbMultiMeasureDimension
-    ) -> List[rdf.qb.MeasureComponentSpecification]:
-        measure_specs: List[rdf.qb.MeasureComponentSpecification] = []
-        for measure in measure_dimension.measures:
-            measure_specs.append(self._get_qb_measure_component_specification(measure))
-
-        return measure_specs
-
-    def _get_qb_measure_component_specification(
-        self, measure: QbMeasure
-    ) -> rdf.qb.MeasureComponentSpecification:
-        if isinstance(measure, ExistingQbMeasure):
-            component_uri = self._new_uri_helper.get_component_uri(
-                get_last_uri_part(measure.measure_uri)
-            )
-            component = rdf.qb.MeasureComponentSpecification(component_uri)
-            component.measure = ExistingResource(measure.measure_uri)
-            component.componentProperties.add(component.measure)
-
-            measure.copy_arbitrary_triple_fragments_to_resources(
-                {RdfSerialisationHint.Component: component}
-            )
-        elif isinstance(measure, NewQbMeasure):
-            component = rdf.qb.MeasureComponentSpecification(
-                self._new_uri_helper.get_component_uri(measure.uri_safe_identifier)
-            )
-            component.measure = rdf.qb.MeasureProperty(
-                self._new_uri_helper.get_measure_uri(measure.uri_safe_identifier)
-            )
-            component.measure.label = measure.label
-            component.measure.comment = measure.description
-            component.measure.subPropertyOf = maybe_existing_resource(
-                measure.parent_measure_uri
-            )
-            component.measure.source = maybe_existing_resource(measure.source_uri)
-            component.measure.range = ExistingResource(self._get_obs_val_data_type())
-            component.componentProperties.add(component.measure)
-
-            measure.copy_arbitrary_triple_fragments_to_resources(
-                {
-                    RdfSerialisationHint.Component: component,
-                    RdfSerialisationHint.Property: component.measure,
-                }
-            )
-
-        else:
-            raise TypeError(f"Unhandled measure type {type(measure)}")
-
-        _logger.debug(
-            "Generated component %s with measure %s.",
-            component.uri,
-            component.measure.uri,
-        )
-
-        return component
-
-    def _get_qb_dimension_specification(
-        self, column_name_uri_safe: str, dimension: QbDimension
-    ) -> rdf.qb.DimensionComponentSpecification:
-        if isinstance(dimension, ExistingQbDimension):
-            component = rdf.qb.DimensionComponentSpecification(
-                self._new_uri_helper.get_component_uri(column_name_uri_safe)
-            )
-            component.dimension = ExistingResource(dimension.dimension_uri)
-            dimension.copy_arbitrary_triple_fragments_to_resources(
-                {RdfSerialisationHint.Component: component}
-            )
-        elif isinstance(dimension, NewQbDimension):
-            component = rdf.qb.DimensionComponentSpecification(
-                self._new_uri_helper.get_component_uri(dimension.uri_safe_identifier)
-            )
-            component.dimension = rdf.qb.DimensionProperty(
-                self._new_uri_helper.get_dimension_uri(dimension.uri_safe_identifier)
-            )
-            component.dimension.label = dimension.label
-            component.dimension.comment = dimension.description
-            component.dimension.subPropertyOf = maybe_existing_resource(
-                dimension.parent_dimension_uri
-            )
-            component.dimension.source = maybe_existing_resource(dimension.source_uri)
-            component.dimension.range = rdfs.Class(
-                self._new_uri_helper.get_class_uri(dimension.uri_safe_identifier)
-            )
-
-            dimension.copy_arbitrary_triple_fragments_to_resources(
-                {
-                    RdfSerialisationHint.Component: component,
-                    RdfSerialisationHint.Property: component.dimension,
-                }
-            )
-
-            if dimension.code_list is not None:
-                component.dimension.code_list = self._get_code_list_resource(
-                    dimension.code_list
-                )
-
-        else:
-            raise TypeError(f"Unhandled dimension component type {type(dimension)}.")
-
-        component.componentProperties.add(component.dimension)
-
-        _logger.debug(
-            "Generated component %s with dimension %s.",
-            component.uri,
-            component.dimension.uri,
-        )
-
-        return component
-
-    def _get_code_list_resource(
-        self, code_list: QbCodeList
-    ) -> Resource[skos.ConceptScheme]:
-        if isinstance(code_list, ExistingQbCodeList):
-            return ExistingResource(code_list.concept_scheme_uri)
-        elif isinstance(code_list, NewQbCodeList):
-            # The resource is created elsewhere. There is a separate CSV-W definition for the code-list
-            return ExistingResource(
-                SkosCodeListNewUriHelper(
-                    code_list, self.cube.uri_style
-                ).get_scheme_uri()
-            )
-        elif isinstance(code_list, NewQbCodeListInCsvW):
-            return ExistingResource(code_list.concept_scheme_uri)
-        else:
-            raise TypeError(f"Unhandled codelist type {type(code_list)}")
-
-    def _get_qb_attribute_specification(
-        self, column_name_uri_safe: str, attribute: QbAttribute
-    ) -> rdf.qb.AttributeComponentSpecification:
-        if isinstance(attribute, ExistingQbAttribute):
-            component = rdf.qb.AttributeComponentSpecification(
-                self._new_uri_helper.get_component_uri(column_name_uri_safe)
-            )
-            if isinstance(attribute, QbAttributeLiteral):
-                component.attribute = ExistingResourceWithLiteral(
-                    attribute.attribute_uri
-                )
-                component.attribute.range = ExistingResource(
-                    get_data_type_uri_from_str(attribute.data_type)
-                )
-            else:
-                component.attribute = ExistingResource(attribute.attribute_uri)
-
-            attribute.copy_arbitrary_triple_fragments_to_resources(
-                {RdfSerialisationHint.Component: component}
-            )
-        elif isinstance(attribute, NewQbAttribute):
-            component = rdf.qb.AttributeComponentSpecification(
-                self._new_uri_helper.get_component_uri(attribute.uri_safe_identifier)
-            )
-            component.attribute = rdf.qb.AttributeProperty(
-                self._new_uri_helper.get_attribute_uri(attribute.uri_safe_identifier)
-            )
-            component.attribute.label = attribute.label
-            component.attribute.comment = attribute.description
-            component.attribute.subPropertyOf = maybe_existing_resource(
-                attribute.parent_attribute_uri
-            )
-            component.attribute.source = maybe_existing_resource(attribute.source_uri)
-            # TODO: Find some way to link the codelist we have to the ComponentProperty?
-            if isinstance(attribute, QbAttributeLiteral):
-                component.attribute.range = ExistingResource(
-                    get_data_type_uri_from_str(attribute.data_type)
-                )
-            attribute.copy_arbitrary_triple_fragments_to_resources(
-                {
-                    RdfSerialisationHint.Component: component,
-                    RdfSerialisationHint.Property: component.attribute,
-                }
-            )
-        else:
-            raise TypeError(f"Unhandled attribute component type {type(attribute)}.")
-
-        component.componentRequired = attribute.is_required
-        component.componentProperties.add(component.attribute)
-
-        _logger.debug(
-            "Generated component %s with attribute %s.",
-            component.uri,
-            component.attribute.uri,
-        )
-
-        return component
-
-    def _get_rdf_file_dependencies(self) -> List[RdfGraphDependency]:
-        """
-        Define RDF dependencies of this CSV-W file which help the end-user understand the triples expressed by this CSV-W.
-
-        :return: RDF resource models to define dependencies on other RDF files.
-        """
-        rdf_file_dependencies: List[RdfGraphDependency] = []
-
-        dimension_columns = get_columns_of_dsd_type(self.cube, NewQbDimension)
-        for column in dimension_columns:
-            dimension = column.structural_definition
-            code_list = dimension.code_list
-            if isinstance(code_list, NewQbCodeList):
-                dependency = RdfGraphDependency(
-                    self._new_uri_helper.get_void_dataset_dependency_uri(
-                        code_list.metadata.uri_safe_identifier
-                    )
-                )
-
-                code_list_writer = self._get_writer_for_code_list(code_list)
-                dependency.uri_space = (
-                    code_list_writer.uri_helper.get_uri_prefix_for_doc()
-                )
-
-                dependency.dependent_rdf_file_location = ExistingResource(
-                    f"./{code_list_writer.csv_metadata_file_name}"
-                )
-
-                rdf_file_dependencies.append(dependency)
-
-        return rdf_file_dependencies
-
     def _get_writer_for_code_list(self, code_list) -> SkosCodeListWriter:
         return SkosCodeListWriter(code_list, self.cube.uri_style)
-
-    def _get_new_attribute_value_resources(self) -> List[NewAttributeValueResource]:
-        """
-        :return: RDF resource models to define New Attribute Values
-        """
-        new_attribute_value_resources: List[NewAttributeValueResource] = []
-        attribute_columns = get_columns_of_dsd_type(self.cube, QbAttribute)
-        for column in attribute_columns:
-            if isinstance(column.structural_definition, NewQbAttribute):
-                column_identifier = column.structural_definition.uri_safe_identifier
-            else:
-                column_identifier = column.uri_safe_identifier
-
-            for value in column.structural_definition.new_attribute_values:  # type: ignore
-                assert isinstance(value, NewQbAttributeValue)
-
-                attribute_value_uri = self._new_uri_helper.get_attribute_value_uri(
-                    column_identifier, value.uri_safe_identifier
-                )
-                new_attribute_value_resource = NewAttributeValueResource(
-                    attribute_value_uri
-                )
-                new_attribute_value_resource.label = value.label
-                new_attribute_value_resource.comment = value.description
-                new_attribute_value_resource.source_uri = maybe_existing_resource(
-                    value.source_uri
-                )
-                new_attribute_value_resource.parent_attribute_value_uri = (
-                    maybe_existing_resource(value.parent_attribute_value_uri)
-                )
-
-                _logger.debug(
-                    "Generated New Attribute Value %s.",
-                    new_attribute_value_resource.uri,
-                )
-
-                new_attribute_value_resources.append(new_attribute_value_resource)
-
-        return new_attribute_value_resources
-
-    def _get_new_unit_resources(self) -> List[NewUnitResource]:
-        """
-        :return: RDF Resources to defined new units
-        """
-        units: Set[QbUnit] = {
-            u
-            for col in get_columns_of_dsd_type(self.cube, QbMultiUnits)
-            for u in col.structural_definition.units  # type: ignore
-        }
-
-        units |= {
-            col.structural_definition.unit
-            for col in get_columns_of_dsd_type(self.cube, QbObservationValue)
-            if col.structural_definition.unit is not None
-        }
-
-        def get_new_base_units(u: QbUnit) -> set[QbUnit]:
-            if (
-                isinstance(u, NewQbUnit)
-                and u.base_unit is not None
-                and isinstance(u.base_unit, NewQbUnit)
-            ):
-                new_base_units = get_new_base_units(u.base_unit)
-                new_base_units.add(u.base_unit)
-
-                return new_base_units
-
-            return set()
-
-        units |= {base_unit for u in units for base_unit in get_new_base_units(u)}
-
-        new_units: Set[NewQbUnit] = {u for u in units if isinstance(u, NewQbUnit)}
-
-        new_unit_resources: List[NewUnitResource] = []
-        for unit in new_units:
-            unit_uri = self._get_unit_uri(unit)
-            new_unit_resource = NewUnitResource(unit_uri)
-            new_unit_resource.label = unit.label
-            new_unit_resource.comment = unit.description
-            new_unit_resource.source_uri = maybe_existing_resource(unit.source_uri)
-
-            maybe_unit_uri = (
-                None if unit.base_unit is None else self._get_unit_uri(unit.base_unit)
-            )
-
-            new_unit_resource.base_unit_uri = maybe_existing_resource(maybe_unit_uri)
-            new_unit_resource.base_unit_scaling_factor = unit.base_unit_scaling_factor
-
-            new_unit_resource.has_qudt_quantity_kind = maybe_existing_resource(
-                unit.qudt_quantity_kind_uri
-            )
-            new_unit_resource.qudt_conversion_multiplier = (
-                unit.si_base_unit_conversion_multiplier
-            )
-
-            _logger.debug("Generated new unit resource %s.", new_unit_resource.uri)
-
-            new_unit_resources.append(new_unit_resource)
-
-        return new_unit_resources
 
     def _generate_csvqb_column(self, column: CsvColumn) -> Dict[str, Any]:
         _logger.debug(
@@ -1006,101 +455,6 @@ class QbWriter(WriterBase):
             )
 
         return csvw_col
-
-    def _get_observation_value_col_for_title(
-        self, col_title: str
-    ) -> QbColumn[QbObservationValue]:
-        """
-        Gets the matching observation value column for the given attributes/units column (if there is one).
-        """
-        _logger.debug(
-            "Getting the matching observation value column for the column with title '%s'",
-            col_title,
-        )
-
-        obs_value_columns = get_columns_of_dsd_type(self.cube, QbObservationValue)
-        obs_columns_for_column = [
-            obs_col
-            for obs_col in obs_value_columns
-            if col_title == obs_col.csv_column_title
-        ]
-
-        num_of_obs_val_cols = len(obs_columns_for_column)
-        _logger.debug(
-            "Found %d observation value column(s) for the column with title '%s'",
-            num_of_obs_val_cols,
-            col_title,
-        )
-        if num_of_obs_val_cols != 1:
-            raise Exception(
-                f'Could not find one observation value column. Found {num_of_obs_val_cols} for title: "{col_title}".'
-            )
-        return obs_columns_for_column[0]
-
-    def _get_about_url_for_csvw_col_in_pivoted_shape_cube(
-        self, column: QbColumn
-    ) -> Optional[str]:
-        obs_val_col: Optional[QbColumn[QbObservationValue]] = None
-        obs_val_cols = get_columns_of_dsd_type(self.cube, QbObservationValue)
-        is_single_measure = len(obs_val_cols) == 1
-        _logger.debug(
-            "Getting about url for column with title '%s'", column.csv_column_title
-        )
-
-        # If the column represents a QbObservationValue, then simply assign the obs_val_column to this column.
-        if isinstance(column.structural_definition, QbObservationValue):
-            _logger.debug("Column is a observation value column")
-            obs_val_col = column
-        # If the column represents an attribute, set the valueUrl using the _get_observation_value_col_for_title function
-        elif isinstance(column.structural_definition, QbAttribute):
-            _logger.debug("Column is a an attribute column")
-            col_title = column.structural_definition.get_observed_value_col_title()
-            obs_val_col = self._get_obs_val_col_for_col_title(
-                col_title, obs_val_cols, is_single_measure
-            )
-        # If the column represents units, set the valueUrl using the _get_observation_value_col_for_title function
-        elif isinstance(column.structural_definition, QbMultiUnits):
-            _logger.debug("Column is a a multi-unit column")
-            col_title = column.structural_definition.observed_value_col_title
-            obs_val_col = self._get_obs_val_col_for_col_title(
-                col_title, obs_val_cols, is_single_measure
-            )
-        else:
-            obs_val_col = None
-
-        return (
-            self._get_observation_uri_for_pivoted_shape_data_set(obs_val_col)
-            if obs_val_col is not None
-            else None
-        )
-
-    def _get_obs_val_col_for_col_title(
-        self,
-        col_title: Optional[str],
-        obs_val_cols: List[QbColumn[QbObservationValue]],
-        is_single_measure: bool,
-    ) -> Optional[QbColumn[QbObservationValue]]:
-        _logger.debug(
-            "Getting observation value column for column title '%s'", col_title
-        )
-        obs_val_col: Optional[QbColumn[QbObservationValue]] = None
-
-        if is_single_measure:
-            obs_val_col = obs_val_cols[0]
-            _logger.debug("The cube is a single measure cube.")
-        else:
-            _logger.debug("The cube is a multi measure cube.")
-            if col_title is not None:
-                obs_val_col = self._get_observation_value_col_for_title(col_title)
-
-        if obs_val_col is not None:
-            _logger.debug(
-                "Observation value column for column title '%s' is '%s'",
-                col_title,
-                obs_val_col.csv_column_title,
-            )
-
-        return obs_val_col
 
     def _determine_whether_the_column_is_required(self, column: QbColumn) -> bool:
         return (
@@ -1131,7 +485,9 @@ class QbWriter(WriterBase):
 
         # If the cube is in pivoted shape, check what the column represents to set the aboutUrl
         if self.is_cube_in_pivoted_shape:
-            about_url = self._get_about_url_for_csvw_col_in_pivoted_shape_cube(column)
+            about_url = self._uris.get_about_url_for_csvw_col_in_pivoted_shape_cube(
+                column
+            )
             _logger.debug(
                 "About url for column with tile '%s' is '%s'",
                 about_url,
@@ -1194,7 +550,7 @@ class QbWriter(WriterBase):
         unit_value_uri: str
         if all_units_new:
             _logger.debug("All units are new; they define the column's valueUrl.")
-            unit_value_uri = self._new_uri_helper.get_unit_uri(column_template_fragment)
+            unit_value_uri = self._uris.get_unit_uri(column_template_fragment)
         elif all_units_existing:
             _logger.debug("All units are existing.")
             unit_value_uri = column_template_fragment
@@ -1231,10 +587,10 @@ class QbWriter(WriterBase):
             ]
         )
 
-        column_template_fragment = self._get_column_uri_template_fragment(column)
+        column_template_fragment = self._uri.get_column_uri_template_fragment(column)
         if all_measures_new:
             _logger.debug("All measures are new; they define the column's valueUrl.")
-            return self._new_uri_helper.get_measure_uri(column_template_fragment)
+            return self._uris.get_measure_uri(column_template_fragment)
         elif all_measures_existing:
             _logger.debug("All measures are existing.")
             if column.csv_column_uri_template is None:
@@ -1280,44 +636,6 @@ class QbWriter(WriterBase):
                 f"Unhandled component type {type(column.structural_definition)}"
             )
 
-    def _get_default_property_value_uris_for_observation_value(
-        self,
-        observation_value: QbObservationValue,
-    ):
-        if observation_value.is_pivoted_shape_observation:
-            assert observation_value.measure is not None
-            _logger.debug(
-                "Single-measure observation value propertyUrl defined by measure %s",
-                observation_value.measure,
-            )
-            return self._get_measure_uri(observation_value.measure), None
-        else:
-            # In the standard shape
-            multi_measure_dimension_col = self._get_single_column_of_type(
-                QbMultiMeasureDimension
-            )
-            _logger.debug(
-                "Multi-measure observation value propertyUrl defined by measure column %s",
-                multi_measure_dimension_col.csv_column_title,
-            )
-
-            measure_uri_template = (
-                self._get_measure_dimension_column_measure_template_uri(
-                    multi_measure_dimension_col
-                )
-            )
-            return measure_uri_template, None
-
-    def _get_single_column_of_type(
-        self, t: Type[QbColumnarDsdType]
-    ) -> QbColumn[QbColumnarDsdType]:
-        cols = get_columns_of_dsd_type(self.cube, t)
-        if len(cols) != 1:
-            raise ValueError(
-                f"Found {len(cols)} columns with component type {t} in cube. Expected 1."
-            )
-        return cols[0]
-
     def _get_default_property_value_uris_for_dimension(
         self, column: QbColumn[QbDimension]
     ) -> Tuple[str, Optional[str]]:
@@ -1354,6 +672,9 @@ class QbWriter(WriterBase):
         attribute = column.structural_definition
         column_uri_fragment = self._get_column_uri_template_fragment(column)
         value_uri = self._get_column_uri_template_fragment(column)
+
+        attribute_uri = self._uris.get_attribute_uri(attribute)
+
         if isinstance(attribute, ExistingQbAttribute):
             if len(attribute.new_attribute_values) > 0:
                 _logger.debug(
@@ -1391,198 +712,43 @@ class QbWriter(WriterBase):
         else:
             raise TypeError(f"Unhandled attribute type {type(attribute)}")
 
-    def _get_column_uri_template_fragment(
-        self, column: CsvColumn, escape_value: bool = False
-    ) -> str:
-        if escape_value:
-            return "{" + csvw_column_name_safe(column.uri_safe_identifier) + "}"
-
-        return "{+" + csvw_column_name_safe(column.uri_safe_identifier) + "}"
-
-    _legacy_external_code_list_pattern = re.compile("^(.*)/concept-scheme/(.*)$")
-    _legacy_dataset_local_code_list_pattern = re.compile("^(.*)#scheme/(.*)$")
-    _csvcubed_code_list_pattern = re.compile(
-        "^(.*)#" + re.escape(SCHEMA_URI_IDENTIFIER) + "$"
-    )
-
-    def _get_default_value_uri_for_code_list_concepts(
-        self, column: CsvColumn, code_list: QbCodeList
-    ) -> str:
-        column_uri_fragment = self._get_column_uri_template_fragment(column)
-
-        if isinstance(code_list, ExistingQbCodeList):
-            legacy_external_match = self._legacy_external_code_list_pattern.match(
-                code_list.concept_scheme_uri
-            )
-            legacy_local_match = self._legacy_dataset_local_code_list_pattern.match(
-                code_list.concept_scheme_uri
-            )
-            csvcubed_match = self._csvcubed_code_list_pattern.match(
-                code_list.concept_scheme_uri
-            )
-            if legacy_external_match:
-                _logger.debug(
-                    "Existing concept scheme URI %s matches legacy family/global style.",
-                    code_list.concept_scheme_uri,
-                )
-                m: re.Match = legacy_external_match
-                # ConceptScheme URI:
-                # http://gss-data.org.uk/def/concept-scheme/{code-list-name}
-                # Concept URI:
-                # http://gss-data.org.uk/def/concept-scheme/{code-list-name}/{notation}
-                return f"{m.group(1)}/concept-scheme/{m.group(2)}/{column_uri_fragment}"
-            elif legacy_local_match:
-                _logger.debug(
-                    "Existing concept scheme URI %s matches legacy dataset-local style.",
-                    code_list.concept_scheme_uri,
-                )
-                m: re.Match = legacy_local_match
-                # ConceptScheme URI:
-                # http://gss-data.org.uk/data/gss_data/{family-name}/{dataset-root-name}#scheme/{code-list-name}
-                # Concept URI:
-                # http://gss-data.org.uk/data/gss_data/{family-name}/{dataset-root-name}#concept/{code-list-name}/{notation}
-                return f"{m.group(1)}#concept/{m.group(2)}/{column_uri_fragment}"
-            elif csvcubed_match:
-                _logger.debug(
-                    "Existing concept scheme URI %s matches csvcubed style.",
-                    code_list.concept_scheme_uri,
-                )
-                m: re.Match = csvcubed_match
-                # ConceptScheme URI:
-                # {code-list-uri}#code-list
-                # Concept URI:
-                # {code-list-uri}#{notation}
-                return f"{m.group(1)}#{column_uri_fragment}"
-            else:
-                _logger.warning(
-                    "Existing code list URI %s does not match expected any known convention.",
-                    code_list.concept_scheme_uri,
-                )
-                # Unexpected code-list URI. Does not match expected conventions.
-                return column_uri_fragment
-        elif isinstance(code_list, NewQbCodeList):
+    def _get_default_property_value_uris_for_observation_value(
+        self,
+        observation_value: QbObservationValue,
+    ):
+        if observation_value.is_pivoted_shape_observation:
+            assert observation_value.measure is not None
             _logger.debug(
-                "valueUrl defined by new dataset-local code list %s",
-                code_list.metadata.title,
+                "Single-measure observation value propertyUrl defined by measure %s",
+                observation_value.measure,
             )
-            return SkosCodeListNewUriHelper(
-                code_list, self.cube.uri_style
-            ).get_concept_uri(column_uri_fragment)
-        elif isinstance(code_list, NewQbCodeListInCsvW):
+            return self._get_measure_uri(observation_value.measure), None
+        else:
+            # In the standard shape
+            multi_measure_dimension_col = self._get_single_column_of_type(
+                QbMultiMeasureDimension
+            )
             _logger.debug(
-                "valueUrl defined by legacy dataset-local code list %s",
-                code_list.concept_scheme_uri,
+                "Multi-measure observation value propertyUrl defined by measure column %s",
+                multi_measure_dimension_col.csv_column_title,
             )
-            return re.sub(
-                r"\{.?notation\}", column_uri_fragment, code_list.concept_template_uri
+
+            measure_uri_template = (
+                self._get_measure_dimension_column_measure_template_uri(
+                    multi_measure_dimension_col
+                )
             )
-        else:
-            raise TypeError(f"Unhandled codelist type {type(code_list)}")
+            return measure_uri_template, None
 
-    def _get_unit_uri(self, unit: QbUnit) -> str:
-        if isinstance(unit, ExistingQbUnit):
-            return unit.unit_uri
-        elif isinstance(unit, NewQbUnit):
-            return self._new_uri_helper.get_unit_uri(unit.uri_safe_identifier)
-        else:
-            raise TypeError(f"Unmatched unit type {type(unit)}")
-
-    def _get_measure_uri(self, measure: QbMeasure) -> str:
-        if isinstance(measure, ExistingQbMeasure):
-            return measure.measure_uri
-        elif isinstance(measure, NewQbMeasure):
-            return self._new_uri_helper.get_measure_uri(measure.uri_safe_identifier)
-        else:
-            raise TypeError(f"Unmatched measure type {type(measure)}")
-
-    def _get_observation_uri_for_standard_shape_data_set(self) -> str:
-        dimension_columns_templates: List[str] = []
-        multi_measure_col_template: Optional[str] = None
-
-        for c in self.cube.columns:
-            if isinstance(c, QbColumn):
-                if isinstance(c.structural_definition, QbDimension):
-                    dimension_columns_templates.append(
-                        f"{{{csvw_column_name_safe(c.uri_safe_identifier)}}}"
-                    )
-                elif isinstance(c.structural_definition, QbMultiMeasureDimension):
-                    multi_measure_col_template = (
-                        f"{{{csvw_column_name_safe(c.uri_safe_identifier)}}}"
-                    )
-        return self._new_uri_helper.get_observation_uri(
-            dimension_columns_templates, multi_measure_col_template
-        )
-
-    def _get_observation_uri_for_pivoted_shape_data_set(
-        self, obs_val_column: QbColumn[QbObservationValue]
-    ) -> str:
-        """
-        Provide the obervation uri for the pivoted shape data set
-        """
-        _logger.debug(
-            "Getting observation uri for the observation value column with title '%s'",
-            obs_val_column.csv_column_title,
-        )
-
-        dimension_columns_templates: List[str] = []
-        for c in self.cube.columns:
-            if isinstance(c, QbColumn):
-                if isinstance(c.structural_definition, QbDimension):
-                    col_template_uri = c.uri_safe_identifier
-                    _logger.debug(
-                        "Detected a dimension column in the cube, hence setting the column template uri '%s'",
-                        col_template_uri,
-                    )
-                    dimension_columns_templates.append(
-                        f"{{{csvw_column_name_safe(col_template_uri)}}}"
-                    )
-
-        obs_val_measure = obs_val_column.structural_definition.measure
-        assert obs_val_measure is not None
-        _logger.debug("Observation value column has a measure")
-
-        if isinstance(obs_val_measure, NewQbMeasure):
-            measure_id = obs_val_measure.uri_safe_identifier
-            _logger.debug("The measure is a new measure with id '%s'", measure_id)
-        elif isinstance(obs_val_measure, ExistingQbMeasure):
-            # Yes, this is absolutely nasty, but what else can we do?
-            measure_id = uri_safe(obs_val_measure.measure_uri)
-            _logger.debug("The measure is an existing measure with id '%s'", measure_id)
-        else:
-            raise TypeError(f"Unhandled QbMeasure type {type(obs_val_measure)}")
-
-        return self._new_uri_helper.get_observation_uri(
-            dimension_columns_templates, measure_id
-        )
-
-    def _get_pivoted_cube_slice_uri(self) -> str:
-        # TODO: Dimensions are currently appended in the order in which the appear in the cube.
-        #       We may want to alter this in the future so that the ordering is from
-        #       least entropic dimension -> most entropic.
-        #       E.g. http://base-uri/observations/male/1996/all-males-1996
-
-        dimension_columns_templates: List[str] = [
-            f"{{{csvw_column_name_safe(c.uri_safe_identifier)}}}"
-            for c in get_columns_of_dsd_type(self.cube, QbDimension)
-        ]
-        _logger.debug(
-            "The cube has %d dimension column templates",
-            len(dimension_columns_templates),
-        )
-
-        return self._new_uri_helper.get_slice_across_measures_uri(
-            dimension_columns_templates
-        )
-
-    def _get_about_url(self) -> str:
-        about_url_template = (
-            self._get_pivoted_cube_slice_uri()
-            if self.is_cube_in_pivoted_shape
-            else self._get_observation_uri_for_standard_shape_data_set()
-        )
-
-        _logger.debug("About url template is %s", about_url_template)
-        return about_url_template
+    def _get_single_column_of_type(
+        self, t: Type[QbColumnarDsdType]
+    ) -> QbColumn[QbColumnarDsdType]:
+        cols = get_columns_of_dsd_type(self.cube, t)
+        if len(cols) != 1:
+            raise ValueError(
+                f"Found {len(cols)} columns with component type {t} in cube. Expected 1."
+            )
+        return cols[0]
 
     def _get_primary_key_columns(self) -> List[str]:
         dimension_columns: Iterable[QbColumn] = itertools.chain(
